0508
전사우클 : 방패치기 우클릭누르면 1초동안 모든피해면역
전사 스킬1 : 폭주 : 공격속도 6초간 증가
전사 스킬2 : 돌진 : 전방으로 돌진함 쿨타임
전사 스킬3 : 검기 발싸 : 전방으로 발싸함

전사가 때릴때 검에붙은 콜라이더가 불안정해서 데미지가 잘 안들어갑니다
그래서 전사안에 AttackArea라는 오브젝트를 추가해서 거기에 웨폰스크립트 넣어서
딜 들어가게 했습니다 검은 장식용으로 써야할것같습니다

궁수우클 : 독뿌리고 백스텝  (몬스터 1초스턴)
궁수 스킬1 : 독화살 : 고정데미지 추가
궁수 스킬2 : 폭탄화살 : 적이맞으면 폭발. 맞은적 스턴1초
궁수 스킬3 : 에너지화살 : 기를 모아서 발사하는컨셉. 1.5초모으고 쏘는게 1차
5초 꽉채우고 쏘는게 2차. 데미지의 차이

법사우클 : 점멸
법사 스킬1 : 3갈래전기구체
법사 스킬2 : 얼음회오리 cc기
법사 스킬3 : 메테오

+++ Assets/02.Scripts/monster/BossSkele.cs
@@ -5,8 +5,8 @@ using UnityEngine.AI;
 
 public class BossSkele : MonoBehaviour
 {
-    public int maxHealth=100; //최대 체력
-    public int curHealth=100; //현재 체력
+    public float maxHealth=100; //최대 체력
+    public float curHealth =100; //현재 체력
     public BoxCollider meleeArea; //몬스터 공격범위
     public bool isChase; //추적중인 상태
     public bool isAttack; //현재 공격중
